requirements.txt
requests
python-dotenv
py-cord
PyNaCl
